addons: only exec surveillancepatch if surveillancestation is installed

+++ surveillancepatch/all/usr/sbin/surveillancepatch.sh
@@ -1,34 +1,38 @@
 #!/usr/bin/env ash
 
 scriptver="23.6.1"
-script=Surveillance
+script=SurveillancePatch
 repo="AuxXxilium/arc-addons"
 
-PATH1="/var/packages/SurveillanceStation/target/lib"
-PATH2="/var/packages/SurveillanceStation/target/scripts"
-SPATCH="/usr/lib"
+if [ -d "/var/packages/SurveillanceStation" ]; then
 
-/usr/syno/bin/synopkg stop SurveillanceStation
-sleep 15
+    PATH1="/var/packages/SurveillanceStation/target/lib"
+    PATH2="/var/packages/SurveillanceStation/target/scripts"
+    SPATCH="/usr/lib"
 
-cp -f ${PATH1}/libssutils.so ${PATH1}/libssutils.so.bak
-rm -f ${PATH1}/libssutils.so
-cp -f ${SPATCH}/libssutils.so ${PATH1}/libssutils.so
-chown SurveillanceStation:SurveillanceStation ${PATH1}/libssutils.so
-chmod 0644 ${PATH1}/libssutils.so
+    /usr/syno/bin/synopkg stop SurveillanceStation
+    sleep 15
 
-cp -f ${PATH2}/S82surveillance.sh ${PATH2}/S82surveillance.sh.bak
-rm -f ${PATH2}/S82surveillance.sh
-cp -f ${SPATCH}/S82surveillance.sh ${PATH2}/S82surveillance.sh
-chown SurveillanceStation:SurveillanceStation ${PATH2}/S82surveillance.sh
-chmod 0755 ${PATH2}/S82surveillance.sh
+    cp -f ${PATH1}/libssutils.so ${PATH1}/libssutils.so.bak
+    rm -f ${PATH1}/libssutils.so
+    cp -f ${SPATCH}/libssutils.so ${PATH1}/libssutils.so
+    chown SurveillanceStation:SurveillanceStation ${PATH1}/libssutils.so
+    chmod 0644 ${PATH1}/libssutils.so
 
-rm -f ${PATH2}/license.sh
-cp -f ${SPATCH}/license.sh ${PATH2}/license.sh 
-chown SurveillanceStation:SurveillanceStation ${PATH2}/license.sh
-chmod 0777 ${PATH2}/license.sh
+    cp -f ${PATH2}/S82surveillance.sh ${PATH2}/S82surveillance.sh.bak
+    rm -f ${PATH2}/S82surveillance.sh
+    cp -f ${SPATCH}/S82surveillance.sh ${PATH2}/S82surveillance.sh
+    chown SurveillanceStation:SurveillanceStation ${PATH2}/S82surveillance.sh
+    chmod 0755 ${PATH2}/S82surveillance.sh
 
-echo -e "Surveillance Patch: Successfull!"
+    rm -f ${PATH2}/license.sh
+    cp -f ${SPATCH}/license.sh ${PATH2}/license.sh 
+    chown SurveillanceStation:SurveillanceStation ${PATH2}/license.sh
+    chmod 0777 ${PATH2}/license.sh
 
-sleep 5
-/usr/syno/bin/synopkg start SurveillanceStation
+    echo -e "Surveillance Patch: Successfull!"
+
+    sleep 5
+    /usr/syno/bin/synopkg start SurveillanceStation
+
+fi
